+++ packages/core/types/index.d.ts
@@ -1,5 +1,5 @@
 export { Addon, Subscriber, SideEffect, Params } from './interface';
-export { Route, RouteDescriptor, LoadModifiers } from './utils/createRoute';
+export { Route, RouteDescriptor, LoadModifiers, LoadFn, PreloadFn } from './utils/createRoute';
 export { Response, RedirectResponse, AnyResponse } from './utils/createResponse';
 export { CuriConfig, ConfigOptions } from './createConfig';
 import createConfig from './createConfig';