+++ packages/core/types/utils/createRoute.d.ts
@@ -6,13 +6,14 @@ export declare type Title = string | ((params?: object, data?: any) => string);
 export declare type LoadFn = (params?: object, location?: HickoryLocation, modifiers?: LoadModifiers, addons?: {
     [key: string]: AddonGet;
 }) => Promise<any>;
+export declare type PreloadFn = () => Promise<any>;
 export interface RouteDescriptor {
     name: string;
     path: string;
     pathOptions?: RegExpOptions;
     body?: () => any;
     children?: Array<RouteDescriptor>;
-    preload?: () => Promise<any>;
+    preload?: PreloadFn;
     load?: LoadFn;
     title?: Title;
     extra?: {
@@ -28,7 +29,7 @@ export interface Route {
     body: () => any;
     getBody: () => any;
     children: Array<Route>;
-    preload: () => Promise<any>;
+    preload: PreloadFn;
     load: LoadFn;
     keys: Array<string | number>;
     match: (pathname: string, rc: ResponseCreator, parentPath?: string) => boolean;