+++ packages/core/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Export `PreloadFn` and `LoadFn` types.
 * Automatically redirect after emitting response.
 
 ## 1.0.0-beta.9